requirements.txt
pillow==10.2.0
pipx==1.4.3
SciPy==1.11.4
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-genai
ultralytics==1.12.0
openCV-Python==
python-dotenv==1.0.0
